requirements.txt
Django==3.2.6
django-appconf==1.0.4
django-cors-headers==3.8.0
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
drf-spectacular==0.18.0
factory-boy==3.2.0
invoke==1.6.0
Pillow==8.3.1
psycopg2==2.9.1
qrcode==7.3.1
Werkzeug==2.0.1